Refactored Menu Section into Client Comp
