+++ menumanagement/src/app/menu/page.js
@@ -2,6 +2,7 @@ import { getMenuFixtures } from '../../lib/getMenuFixtures';
 import { categories } from '../../lib/categoryFixtures';
 import { CategoryItems } from '../../components/CategoryItems';
 import { ItemCard } from '../../components/ItemCard';
+import OrderingApp from '../../components/OrderingApp'
 import MenuLink from '../../components/MenuLink';
 import styles from './page.module.css';
 import Link from 'next/link';
@@ -15,22 +16,8 @@ const Menu = async () => {
   return (
     <div className={styles.menu}>
       <h1> Koo Koo Chicken Menu </h1>
-      <a href='#Specialty'>Specialties</a>
       <MenuLink categories={categories} />
-      {categories.map((category, i) => {
-        const itemsByCategory = items.filter(item => {
-          return item.category === category;
-        })
-
-        return (
-          <div id={category} className={styles.categorySections}>
-            <div>
-              <h2>{category}</h2>
-            </div>
-            <CategoryItems items={itemsByCategory} />
-          </div>
-        )}
-      )}
+      <OrderingApp items={items}/>
     </div>
   )
 }
