+++ menumanagement/src/app/menu/page.module.css
@@ -1,19 +1,24 @@
-.categorySections {
-  /* justify-content: center; */
-  /* border-color: rgba(69, 69, 206, 0.587);
-  border-style: solid; */
-  /* height: 100vh;
-  width: 80vh; */
-  /* display: flex; */
-  /* flex-direction: column; */
-  /* flex-wrap: wrap; */
+/* .categorySections {
+  justify-content: center;
+  border-color: rgba(69, 69, 206, 0.587);
+  border-style: solid;
+  height: 100vh;
+  width: 80vh;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
   margin: 0.5em;
   padding: 0.1em;
   justify-content: center;
   align-items: center;
-}
+} */
 
 .menu {
   flex-direction: column;
   justify-content: center;
+  top:30px;
+}
+
+.menuSection {
+  top: 30px;
 }
