+++ menumanagement/src/components/MenuLink.js
@@ -7,17 +7,23 @@ import Link from 'next/link';
 
 const MenuLink = ({categories}) => {
   return (
-    <div className={styles.linkBar}>
-      {categories.map((category, i) => {
-        return (
-          <a href={`#${category}`} key={`link-${category}-${i}`}>
-            <span className={styles.inLine}>
-              {category}
-            </span>
-          </a>
-        )
-      }
-      )}
+    <div className={styles.stickyBar}>
+      <div>
+
+
+      <div className={styles.linkBar}>
+        {categories.map((category, i) => {
+          return (
+            <a href={`#${category}`} key={`link-${category}-${i}`}>
+              <div className={styles.inLine}>
+                {category}
+              </div>
+            </a>
+          )
+        }
+        )}
+      </div>
+      </div>
     </div>
   )
 }
