+++ menumanagement/src/styles/CategoryItems.module.css
@@ -16,4 +16,5 @@
   display: grid;
   grid-template-columns: repeat(2, calc(50% - .5em));
   grid-gap: .5em .5em;
+  /* top: 30px; */
 }
