+++ menumanagement/src/styles/MenuLink.module.css
@@ -2,16 +2,29 @@
   scroll-margin-top: 4rem;
 }
 
+/* .stickyBar {
+  z-index: 20;
+  position: fixed;
+  top: 12px;
+  background-color: aqua;
+  margin: 0 .2em 0 .2em;
+  height: 4vh
+} */
+
 .linkBar {
   display: flex;
   flex-direction: row;
+  flex: 1 1 0%;
+  /* flex-wrap: nowrap; */
   justify-content: space-between;
   overflow-x: auto;
   -ms-overflow-style: none;  /* hide scrollbar for IE and Edge */
   scrollbar-width: none;  /* hide scrollbar for Firefox */
   /* overflow-x: hidden; */
-  max-height: 2em;
-  z-index: 100;
+  height: 3vh;
+  /* z-index: 100;
+  position: sticky;
+  top: 0px; */
   /* position: sticky; hidden b/c causes flashing before scrolling to content */
 }
 
@@ -20,6 +33,7 @@
   padding: .2em;
   min-width: 100%;
   max-width: 100%;
+  height: 2vh;
   margin: 0.1em;
   white-space: nowrap;
   scroll-margin-top: 4rem;
