+++ menumanagement/src/styles/OrderingApp.module.css
@@ -1,0 +1,18 @@
+.categorySections {
+  /* justify-content: center; */
+  /* border-color: rgba(69, 69, 206, 0.587);
+  border-style: solid; */
+  /* height: 100vh;
+  width: 80vh; */
+  /* display: flex; */
+  /* flex-direction: column; */
+  /* flex-wrap: wrap; */
+  margin: 0.5em;
+  padding: 0.1em;
+  justify-content: center;
+  align-items: center;
+}
+
+.menuSection {
+  top: 30px;
+}
